structure accounts dir, create new init migration
