+++ backend/src/adapters/postgres/models/__init__.py
@@ -6,7 +6,8 @@ from .accounts import (
     PasswordResetTokenModel,
 )
 from .base import Base
-from .enums import GenderEnum
+from .enums import GenderEnum, LanguageEnum
+from .snippets import SnippetModel, TagModel
 
 __all__ = [
     "Base",
@@ -16,4 +17,7 @@ __all__ = [
     "ActivationTokenModel",
     "PasswordResetTokenModel",
     "GenderEnum",
+    "LanguageEnum",
+    "SnippetModel",
+    "TagModel",
 ]
