+++ backend/src/adapters/postgres/models/accounts/__init__.py
@@ -1,0 +1,8 @@
+from .profile import UserProfileModel
+from .tokens import (
+    TokenBaseModel,
+    ActivationTokenModel,
+    PasswordResetTokenModel,
+    RefreshTokenModel,
+)
+from .user import UserModel
