+++ backend/src/adapters/postgres/models/enums.py
@@ -4,3 +4,8 @@ from enum import Enum
 class GenderEnum(Enum):
     MALE = "male"
     FEMALE = "female"
+
+
+class LanguageEnum(Enum):
+    PYTHON = "python"
+    JAVASCRIPT = "javascript"
